% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpansion_CEA_TH_NCT.R
\name{coexp.sample.pars}
\alias{coexp.sample.pars}
\title{Internal function of the sim.coexp function}
\usage{
coexp.sample.pars(nruns, var.zeta, coexp.prior, buffer, Ne.prior, NeA.prior,
  time.prior, gene.prior)
}
\description{
sample the parameters of the models from prior distributions.
}
