% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpansion_PT.R
\name{coexp.sample.pars2}
\alias{coexp.sample.pars2}
\title{Internal function of the sim.coexp2 function}
\usage{
coexp.sample.pars2(nruns, var.zeta, coexp.prior, buffer, Ne.prior, NeA.prior,
  time.prior, gene.prior)
}
\description{
sample the parameters of the models from prior distributions.
}
