% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coexpansion_CEA_TH_NCT.R
\name{coexp.sumstat}
\alias{coexp.sumstat}
\title{internal function of the coexpansion model}
\usage{
coexp.sumstat(ms.output, gene.prior)
}
\description{
calculates summary statistics on ms-like files.
}
