% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pop_demog_test.R
\name{demog.sample.pars}
\alias{demog.sample.pars}
\title{internal function of the test.demog function}
\usage{
demog.sample.pars(nruns, var.zeta, coexp.prior, buffer, Ne.prior, NeA.prior,
  time.prior, gene.prior)
}
\description{
internal function of the test.demog function
}
