% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.tools.R
\name{get.prior.table}
\alias{get.prior.table}
\title{Get a table with prior distributions}
\usage{
get.prior.table(model)
}
\arguments{
\item{model}{A model object.}
}
\value{
A data frame with 4 columns: parameter name, first parameter of prior distribution, second parameter of prior distribution, prior distribution.
}
\description{
This function makes a table with the parameters of the model and respective prior distribution.
}
\author{
Marcelo Gehara
}
