% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.sim.R
\name{get.tree.info}
\alias{get.tree.info}
\title{Get tree information}
\usage{
get.tree.info(tree)
}
\arguments{
\item{tree}{The bifurcating tree topology in newick format.}
}
\value{
A list containing the tip names, the newick tree with numbers and a matrix with the node hwighs.
}
\description{
Get information of tip names, associated numbers and node heights.
}
