% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hap.div.R
\name{H.div}
\alias{H.div}
\title{Haplotype Diversity}
\usage{
H.div(x)
}
\arguments{
\item{x}{a DNAbin object}
}
\value{
Number of haplotypes and haplotype diversity and of x.
}
\description{
This function calculates haplotipe diversity from DNAbin sequence file
}
\note{
requires Pegas package
}
\references{
Nei, M., & Tajima, F. (1981). DNA polymorphism detectable by restriction endonucleases. Genetics, 97, 145–163.
}
\author{
Marcelo Gehara
}
