% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iPyrad.alleles.loci2fasta.R
\name{iPyrad.alleles.loci2fasta}
\alias{iPyrad.alleles.loci2fasta}
\title{iPyrad alleles.loci file to fasta alignments conversion}
\usage{
iPyrad.alleles.loci2fasta(alleles.loci, output.dir)
}
\arguments{
\item{alleles.loci}{The name of the alleles.loci file as a character string. The alleles.loci file must be in the working directory.}

\item{output.dir}{The path to the output directory to write the fasta alignments.}
}
\value{
Writes one fasta alignment per locus in the output.dir
}
\description{
Converts iPyrad alleles.loci file to fasta alignmets.
             The alleles.loci is a specific iPyrad format that contains the phased alleles for all loci.
             To get this file use the * option in the output_formats parameter of iPyrad.
}
\author{
Marcelo Gehara (Thanks Gustavo Cabanne for the help)
}
