% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_classifier.R
\name{load.classifier}
\alias{load.classifier}
\title{Load a classifier produced by `save.classifier()`}
\usage{
load.classifier(path, device = NULL)
}
\arguments{
\item{path}{path written by `save.classifier()`}

\item{device}{"cpu" or "cuda"; default chooses cuda when available}
}
\description{
Load a classifier produced by `save.classifier()`
}
