% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.commander.snp.R
\name{ms.commander.snp}
\alias{ms.commander.snp}
\title{Sample parameters and generates msABC string}
\usage{
ms.commander.snp(model, use.alpha = use.alpha, msABC)
}
\arguments{
\item{model}{A model object}

\item{use.alpha}{Logical.If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes.}

\item{msABC}{path to the msABC binary.}
}
\value{
a list with msABC command and sampled parameters.
}
\description{
this function sample parameters from priors and trasform them to coalescent scale to generate a msABC string from a model object generated by the Model Builder, main.menu() function.
}
\note{
This function is used internally for the sim.snp.sumstat function. One my want to run this function to check the msABC string.
}
