% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.commander2.R
\name{ms.commander2}
\alias{ms.commander2}
\title{Sample parameters and generates ms string}
\usage{
ms.commander2(model, use.alpha = use.alpha)
}
\arguments{
\item{model}{A model object}

\item{use.alpha}{Logical.If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes.}
}
\value{
a list with msABC command and sampled parameters.
}
\description{
this function sample parameters from priors and trasform them to coalescent scale to generate a ms command from a model object generated by the Model Builder, main.menu() function.
}
\note{
This function is used internally for the sim.snp.sumstat function. One my want to run this function to check the ms string.
}
