% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msABC.commander.R
\name{msABC.commander}
\alias{msABC.commander}
\title{Sample parameters and generates msABC string}
\usage{
msABC.commander(model, use.alpha = use.alpha)
}
\arguments{
\item{model}{A model object.}

\item{use.alpha}{Logical.If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes.}
}
\value{
a list with msABC command and sampled parameters.
}
\description{
this function sample parameters from priors and trasform them to coalescent scale to generate a msABC command from a model object generated by the Model Builder, main.menu() function.
}
\note{
This function is used internally for the sim.msABC function. One mey want to run this function to check the msABC string.
}
