% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.sumstat.ngs.R
\name{obs.sumstat.ngs}
\alias{obs.sumstat.ngs}
\title{Observed summary statistics for nexgen data}
\usage{
obs.sumstat.ngs(
  model = NULL,
  path.to.fasta = NULL,
  path.to.phylip = NULL,
  pop.assign,
  moments = TRUE,
  ncores = 1
)
}
\arguments{
\item{model}{A model object built by the main.menu.gui() function. Any model with the same number of populations of your empirical data will work. This is just to build the sumstats names correctly.}

\item{path.to.fasta}{Path to the folder containing all fastas to be included in the calculation.
Invariable sites must be included in the fasta alignments.
Invariable loci must also be included. Alignments must contain phased data.
Either path.to.fasta or path.to.phylip must be provided.}

\item{path.to.phylip}{Path to a multi-locus sequential PHYLIP file. Each locus block starts
with a "ntax nchar" header followed by ntax sequence lines. Alternative
to path.to.fasta for datasets with many loci.}

\item{pop.assign}{A two-column data frame with sample names in the first column and the corresponding population membership preferably as numbers in the second column (If you have a single population the numbers won't matter). The numbers should match the population number in the model object.}

\item{moments}{Logical. If TRUE computes the four moments (mean, variance, skewness, kurtosis) of each summary statistic across loci. Default is TRUE.}

\item{ncores}{Number of cores for parallel execution. When ncores > 1, loci are processed in parallel using mclapply. Default is 1.}
}
\value{
A table containing the observed summary stats.
}
\description{
This function calculates the observed summary statistics from an empirical data.
             This summary statistics are the same as those simulated by the sim.sumstat function.
             It is optimized for nexgen data.
}
\note{
This function does not work on Windows systems.
}
\author{
Marcelo Gehara
}
