% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.sim.R
\name{observed.pw.distances}
\alias{observed.pw.distances}
\title{Observed pairwise distances between tips of tree}
\usage{
observed.pw.distances(tree, path)
}
\arguments{
\item{tree}{The species tree topology in newick format.}

\item{path}{path to the directory containing fasta files to be included in the calculation.}
}
\value{
Pairwise distances betwen tips of species tree.
}
\description{
Calculation of paiwise distance between the tips of the tree.
}
