% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.sumstat.R
\name{observed.sumstat}
\alias{observed.sumstat}
\title{Observed summary statistics}
\usage{
observed.sumstat(
  model,
  path.to.fasta,
  fasta.files = list.files(),
  overall.SS = T,
  perpop.SS = T,
  get.moments = T
)
}
\arguments{
\item{model}{A model object built by the main.menu.gui() function.}

\item{path.to.fasta}{Path to the folder containing all fastas to be included in the calculation.}

\item{list.files}{A list of fasta files to be included in the calculation. By default all fastas present in the folder are included.}

\item{moments}{Logical. If TRUE computes the four moments (mean, variance, kurtosis, skewness) of each summary statistics across loci. If False only mean is computed. Defalt is FALSE.}

\item{pop.assign}{A two-column data frame with sample names in the first column and the corresponding population membership as numbers in the second column. The numbers should match the population number in the model object.}

\item{msABC.call}{String. Path to the msABC executable. msABC binaries for Mac's and Linux are included in the package and should work in most computers.
There is no need to change that unless you want to compile the program yourself and point the function to it.}
}
\value{
A list of vectors containing the observed summary stats.
}
\description{
This function calculates the observed summary statistics from an empirical data.
             This summary statistics are the same as those simulated by the sim.sumstat function.
             It is optimized for sanger data.
}
\note{
This function does not work on Windows systems.
}
\author{
Marcelo Gehara
}
