% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.tools.R
\name{plot.priors}
\alias{plot.priors}
\title{Plot prior distribution}
\usage{
\method{plot}{priors}(model, nsamples = 1000, mu.rates = NULL)
}
\arguments{
\item{model}{A model object.}

\item{nsamples}{Number of samples to draw from each prior distribution.}

\item{mu.rates}{List. Distribution to sample the mutation rates. If NULL, rates are sampled from the model. See sim.sumstat for details.}
}
\value{
Graphic
}
\description{
This function plots a density of the simulated prior.
}
\author{
Marcelo Gehara
}
