% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.tools.R
\name{plot.sim.obs}
\alias{plot.sim.obs}
\title{Plot simulated data against observed}
\usage{
\method{plot}{sim.obs}(sim, obs)
}
\arguments{
\item{sim}{A data frame with simulated data.}

\item{obs}{A vector of observed summary stats corresponding to the sim object.}
}
\value{
Graphic
}
\description{
This function plots an histogram of the simulated summary statistics with the corresponding observed value as a red line for comparison.
}
\author{
Marcelo Gehara
}
