% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popassign-data.R
\docType{data}
\name{popassign}
\alias{popassign}
\title{Population assignment file for Dermatonotus muelleri sequences (Deprecated)}
\format{
A data frame with 88 rows and 2 columns (\code{sample}, \code{pop}).
}
\usage{
data(popassign, package = "PipeMaster")
}
\description{
\strong{[Deprecated]}
This dataset is deprecated and will be replaced in a future version.
Population assignment table mapping samples to populations.
}
\references{
Gehara et al. (in review)
}
\keyword{datasets}
\keyword{internal}
