% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_h_menu.R
\name{print.hModel}
\alias{print.hModel}
\title{Print method for hModel objects}
\usage{
\method{print}{hModel}(x, ...)
}
\arguments{
\item{x}{An hModel object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for hModel objects
}
