% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_classifier.R
\name{save.classifier}
\alias{save.classifier}
\title{Save a classifier produced by `tune.nn.classify()`}
\usage{
save.classifier(classifier, path)
}
\arguments{
\item{classifier}{the object returned by `tune.nn.classify()`}

\item{path}{file path (`.pt` extension recommended)}
}
\description{
Save a classifier produced by `tune.nn.classify()`
}
