% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs-data.R
\docType{data}
\name{seqs}
\alias{seqs}
\title{200 sequences of the Dermatonotus muelleri frog (Deprecated)}
\format{
A list of 200 \code{DNAbin} matrices.
}
\usage{
data(seqs, package = "PipeMaster")
}
\description{
\strong{[Deprecated]}
This dataset is deprecated and will be replaced in a future version.
200 sequences stored in a list of DNAbin matrices.
}
\references{
Gehara et al. (in review)
}
\keyword{datasets}
\keyword{internal}
