% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.coaltrees.R
\name{sim.coaltrees}
\alias{sim.coaltrees}
\title{Simulate coalescent trees using ms}
\usage{
sim.coaltrees(
  model,
  nsim.blocks,
  use.alpha = F,
  path = getwd(),
  append.sims = F,
  sim.block.size = 1000
)
}
\arguments{
\item{model}{A model object built by the main.menu.gui() function.}

\item{nsim.blocks}{Number of blocks to simulate. The total number of simulations is: nsim.blocks x sim.block.size.}

\item{use.alpha}{Logical.If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes. Default is FALSE.}

\item{path}{Path to write the output. By default outputs will be saved in the working directory.}

\item{append.sims}{Logical. If TRUE simulations will be appended in the last output. Default is FALSE.}

\item{sim.block.size}{Simulations are performed in blocks. This argument defines the size of the block in number of simulations, i.e. how many simulations to run per block.
A block of 1000 will work for most cases. Increse the total number of simulations with nsim.block argument.}
}
\value{
Writes trees and parameters to the path directory.
}
\description{
This function simulates coalescent trees and writes them into a file.
}
\references{
Hudson R.R. (2002) Generating samples under a Wright-Fisher neutral model of genetic variation. Bioinformatics, 18, 337–338.
}
\author{
Marcelo Gehara
}
