% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pop_demog_test.R
\name{sim.demog}
\alias{sim.demog}
\title{internal function}
\usage{
sim.demog(nsims, coexp.prior, Ne.prior, NeA.prior, time.prior, gene.prior,
  alpha = alpha, append.sims = F, path = getwd())
}
\description{
internal function
}
