% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.sumstat.R
\name{sim.ms.sumstat}
\alias{sim.ms.sumstat}
\title{Simulate summary statistics using ms and PopGenome}
\usage{
sim.ms.sumstat(
  model,
  use.alpha = F,
  nsim.blocks = 1,
  sim.block.size = 1000,
  path = getwd(),
  perpop.SS = T,
  overall.SS = T,
  output.name = "model",
  get.moments = F,
  append.sims = F
)
}
\arguments{
\item{model}{A model object built by the main.menu.gui() function.}

\item{use.alpha}{Logical.If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes. Default is FALSE.}

\item{nsim.blocks}{Number of blocks to simulate. The total number of simulations is: nsim.blocks x sim.block.size.}

\item{sim.block.size}{Simulations are performed in blocks. This argument defines the size of the block in number of simulations, i.e. how many simulations to run per block.
A block of 1000 will work for most cases. Increse the total number of simulations with nsim.block argument.}

\item{path}{Path to write the output. By default outputs will be saved in the working directory.}

\item{perpop.SS}{Logical. If TRUE calculates the summary statistics per population. Default is TRUE.}

\item{overall.SS}{Logical. If TRUE calculates the summary statistics across all your populations. Defaut is FALSE.}

\item{output.name}{String. The prefix of the output names. Defalt is "model"}

\item{get.moments}{Logical. If TRUE computes the four moments (mean, variance, kurtosis, skewness) of each summary statistics across loci. If False only mean is computed. Defalt is FALSE.}

\item{append.sims}{Logical. If TRUE simulations will be appended in the last output. Default is FALSE.}
}
\value{
Writes simulations and parameters to the path directory.
}
\description{
Simulate summary statistics using ms and PopGenome
}
\references{
Hudson R.R. (2002) Generating samples under a Wright-Fisher neutral model of genetic variation. Bioinformatics, 18, 337–338.

Pfeifer, B. et al. (2014) PopGenome: An Efficient Swiss Army Knife for Population Genomic Analyses in R. Mol Biol Evol 31(7): 1929-1936.<doi:10.1093/molbev/msu136>
}
\author{
Marcelo Gehara
}
