% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.sim.R
\name{sim.sp.tree}
\alias{sim.sp.tree}
\title{Simulation of phylogenetic networks under the coalescent}
\usage{
sim.sp.tree(tree, Ne.prior, bifurcating = T, migration = F, admixture = F,
  mig, hib.clade, hib.priors, major.sister, minor.sister, bp, mi, nsims, nloci,
  gen.time, time.modif, time.scalar = 1)
}
\arguments{
\item{tree}{The bifurcating tree topology in newick format.}

\item{Ne.prior}{A vector of two values representing the min and max boundaries of a uniform prior for Ne.}

\item{bifurcating}{Logical. If TRUE the bifurcating topology is simulated.}

\item{migration}{Logical. If TRUE gene flow between two branches is allowed. Migrarion prior need to be specified in the 'mig' argument.}

\item{admixture}{Logical. If TRUE an admixture event between two branches is allowed.}

\item{mig}{A vector of two values representing the min and max boundaries of a uniform prior for gene flow in number of migrant coppies (2Nm).}

\item{hib.clade}{a vector of numbers indicating the hybrid clade. Number associated with terminals can be checked with the get.tree.info function.}

\item{hib.priors}{a vector of 5 numbers representing the lower, upper boundaries of the hybridization time; the upper boundary of the minor sister; lower and upper boundaries of admixture proportion (0-1).}

\item{major.sister}{vector of numbers indicating the the major sister clade. Number associated with terminals can be checked with the get.tree.info function.}

\item{minor.sister}{vector of numbers indicating the minor sister clade. Number associated with terminals can be checked with the get.tree.info function.}

\item{bp}{a vector of two numbers indicsting the mean and SD of base pairs across all loci.}

\item{mi}{a vector of two numbers indicsting the min and max mutation rate across all loci.}

\item{nsims}{Total number of simulations.}

\item{nloci}{Number of loci to be simulated in each iteration.}

\item{gen.time}{Generation time in years.}

\item{time.modif}{A time modifier to alter the age of the nodes in the newick tree.}

\item{time.scalar}{multiplier for three heights in the newick tree.}
}
\description{
Simulation of bifurcating and nonbifurcating species tree
}
