% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.tools.R
\name{update.priors}
\alias{update.priors}
\title{Update priors using a prior table}
\usage{
\method{update}{priors}(tab, model)
}
\arguments{
\item{tab}{A prior table generated by the get.prior.table function.}

\item{model}{A model object.}
}
\value{
A data frame with 4 columns: parameter name, first parameter of prior distribution, secound parameter of prior distribution, prior distribution.
}
\description{
This function updates the prior of your models using a prior table. see get.prior.table function
}
\author{
Marcelo Gehara
}
