% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta.snps.2ms.R
\name{fasta.snp.2ms}
\alias{fasta.snp.2ms}
\title{internal function of the obs.snp.sumstat function}
\usage{
fasta.snp.2ms(path.to.fasta, fasta.files, write.file = T, pop.assign)
}
\description{
transforms fasta alignments in ms-like.
}
