% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.data.structure.R
\name{get.data.structure}
\alias{get.data.structure}
\title{Read the observed data to get the simulation parameters; base pairs and number of individuals per population.}
\usage{
get.data.structure(model, path.to.fasta, pop.assign, sanger = F)
}
\arguments{
\item{model}{A model object generated with main.menu function.}

\item{path.to.fasta}{path to the fasta alignments directory.}

\item{pop.assign}{a two column data frame with poppulation assignemt of individuals. First column must have the samples name, the secound column the population in numbers, 1 for pop1, 2 for pop2 etc.}

\item{sanger}{logical. If TRUE the inheritance scalar and mutation rates set up in the main.menu are kept. If FALSE the mutation rate is propagated to all loci. Defaut is FALSE.}
}
\value{
Model object with updated gene parameters.
}
\description{
Read the observed data to get the simulation parameters; base pairs and number of individuals per population.
}
\author{
Marcelo Gehara
}
