% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MENU_main.menu.R
\name{main.menu}
\alias{main.menu}
\title{Model Builder}
\usage{
main.menu(input = NULL, ms.string = NULL)
}
\arguments{
\item{input}{A model object to be used as template. Can be left blank, defalt NULL.}

\item{ms.string}{a character string describing the model. This can be obtained by running}
}
\value{
An object representing a diversification model.
}
\description{
This function starts the menu for model building.
}
\examples{
my.model<-main.menu()
sim.sumstat(my.model)
}
