% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observed.sumstat.ngs.R
\name{obs.sumstat.ngs}
\alias{obs.sumstat.ngs}
\title{Observed summary statistics for nexgen data}
\usage{
obs.sumstat.ngs(model = NULL, path.to.fasta, pop.assign, moments = F,
  msABC.call = get.msABC())
}
\arguments{
\item{model}{A model object bult by the main.menu function. Any model with the same number of populations of your empirical data will work. This is just to build the sumstats names correctly.}

\item{path.to.fasta}{Path to the folder containing all fastas to be included in the calculation.}

\item{pop.assign}{A two-column data frame with sample names in the first column and the corresponding population membership preferably as numbers in the second column (If you have a single population the numbers wont matter). The numbers should match the population number in the model object.}

\item{moments}{Logical. If TRUE computes the four moments (mean, variance, kurtosis, skewness) of each summary statistics across loci. If False only mean is computed. The defalt is FALSE.}

\item{msABC.call}{String. Path to the msABC executable. msABC binaries for Mac's and Linux are included in the package and should work in most computers.
There is no need to change that unless you want to compile the program yourself and point the function to it.}
}
\value{
A list of vectors containing the observed summary stats.
}
\description{
This function calculates the observed summary statistics from an empirical data. This summary statistics are the same as those simulated by the sim.snp.sumstat function. It is optimized for nexgen data.
}
\note{
This function does not work on Windows systems.
}
\author{
Marcelo Gehara
}
