% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeMaster.R
\docType{package}
\name{PipeMaster}
\alias{PipeMaster}
\alias{PipeMaster-package}
\title{PipeMaster: A package to build and simulate coalescent models}
\description{
PipeMaster: A package to build and simulate coalescent models
}
\section{PipeMaster functions}{

observed.snp.sumstat
}

