% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipeMaster.R
\docType{package}
\name{PipeMaster}
\alias{PipeMaster-package}
\alias{PipeMaster}
\title{PipeMaster: A package to build and simulate coalescent models}
\description{
PipeMaster helps you build and simulate coalescent models within R. One can use the interactive menu to build a colescent model that can be further simulated. It can also simulate simutaineous expansion for single locus data across multiple species. It can output summary statistics and coalescent trees. It will output hyper summary statistics in the case of coexpansion models. It can also calculate the same summary stats from your own data. It uses the ms coalescent simulator implemented as a function in "phyclust" and a built-in msABC C implementation to calculate summary stats.
}
\section{PipeMaster functions}{

observed.snp.sumstat
}

\seealso{
Useful links:
\itemize{
  \item \url{http://www.marcelogehara.org/}
  \item Report bugs at \url{https://github.com/gehara/PipeMaster}
}

}
\author{
\strong{Maintainer}: Marcelo Gehara \email{marcelo.gehara@gmail.com}

Authors:
\itemize{
  \item Guilherme Mazzocchini
  \item Frank Burbrink
}

}
