% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.tools.R
\name{PlotModel}
\alias{PlotModel}
\title{Plot model}
\usage{
PlotModel(
  model,
  use.alpha = F,
  average.of.priors = F,
  xlab = "Population",
  ylab = "Time before present (generations)",
  axes = T
)
}
\arguments{
\item{model}{A model object generated in the main.menu.}

\item{use.alpha}{Logical. If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes. Default is FALSE.
This argument changes ONLY the MOST RECENT demographic change. You should indicate the population numbers for the exponential change
together with the logical argument in a vector. Ex.c(T,1,2)}

\item{average.of.priors}{logical. if TRUE parameters for the plot will be equal to the average of prior values.}

\item{xlab}{Character. Label for the x-axis. Default is "Population".}

\item{ylab}{Character. Label for the y-axis. Default is "Time before present (generations)".}

\item{axes}{logical. if TRUE plot axes with models. Default is TRUE.}
}
\value{
Graphic
}
\description{
This function plots a graphical representation of your model.
}
\author{
Marcelo Gehara. This function is a wrapper of the PlotMS function of the POPDemog package.
}
