% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.tools.R
\name{PlotModel}
\alias{PlotModel}
\title{Plot model}
\usage{
PlotModel(model, use.alpha = F)
}
\arguments{
\item{model}{A model object generated in the main.menu.}

\item{use.alpha}{logical. TRUE for exponential pop changes. FALSE fro sudden changes. Default is FALSE.}
}
\value{
Graphic
}
\description{
This function plots a graphical representation of your model.
}
\author{
Marcelo Gehara. This function is a wrapper of the PlotMS function of the POPDemog package.
}
