% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.tools.R
\name{PlotModel}
\alias{PlotModel}
\title{Plot model}
\usage{
PlotModel(model, use.alpha = F)
}
\arguments{
\item{model}{A model object generated in the main.menu.}

\item{use.alpha}{Logical. If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes. Default is FALSE.
This argument changes ONLY the MOST RECENT demographich change.}
}
\value{
Graphic
}
\description{
This function plots a graphical representation of your model.
}
\author{
Marcelo Gehara. This function is a wrapper of the PlotMS function of the POPDemog package.
}
