% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popassign-data.R
\docType{data}
\name{popassign}
\alias{popassign}
\title{assignment file example for the Dermatonotus muelleri sequences}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(popassign, package = "PipeMaster")
}
\description{
assignment file
}
\references{
Gehara et al. (in review)
}
\keyword{datasets}
