% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.planner.R
\name{PopPlanner}
\alias{PopPlanner}
\title{Calls PopPlanner, a java application with GUI to build ms population models generating the corresponding ms string.}
\usage{
PopPlanner()
}
\description{
\lifecycle{deprecated}

`PopPlanner()` is deprecated. Use [main.menu.gui()] instead to build
demographic models interactively.

When you run this function PopPlanner will pop up. PopPlanner has an easy to use interface to build diversification models.
The resulting ms string will be generated in the bottom of the PopPlanner window. This can be used as an input to the main.menu.gui() function.
}
