% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.planner.R
\name{PopPlanner}
\alias{PopPlanner}
\title{Calls PopPlanner, a java aplication with GUI to build ms population models generating the coresponding ms string.}
\usage{
PopPlanner()
}
\description{
When you run this function PopPlanner will pop up. PopPlanner has a easy to use interface to build a diversification models.
           The resulting ms string will be generated in the botton of the PopPlanner window. This can be used as an input to the main.menu function.
}
