% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_conversion_functions.R
\name{read.phylip.loci}
\alias{read.phylip.loci}
\title{Converts ms output to DNAbin file}
\usage{
read.phylip.loci(filepath)
}
\arguments{
\item{ms.output}{A list of strings representing a ms output.}

\item{bp.length}{The number of base pairs used to calculate theta for the ms simulation.}
}
\value{
a DNAbin object
}
\description{
This function will take a ms-like output and convert it to DNAbin format.
}
\note{
This function is used internally by all the main functions used to simulate coexpantion models.
One can read in an ms output with readLines().
}
\examples{
# theta = 4Ne x mi x bp
Ne = 100000 # effective pop size
mi = 1e-8  # per base pair per generation mutation rate
bp = 1000  # number of base pairs
theta<-4*Ne*mi*bp
x<-ms(nsam=10, nrep=1, opts = paste("-t",theta))
y<-ms.to.DNAbin(x,bp=1000)
nuc.div(y)
}
\author{
Marcelo Gehara
}
