% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqs-data.R
\docType{data}
\name{seqs}
\alias{seqs}
\title{200 sequences of the Dermatonotus muelleri frog}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\usage{
data(200seqs)
}
\description{
200 sequences stored in a list.
}
\references{
Gehara et al. (in review)
}
\keyword{datasets}
