% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.msABC.R
\name{sim.msABC.sanger}
\alias{sim.msABC.sanger}
\title{Simulate summary statistics using msABC}
\usage{
sim.msABC.sanger(
  model,
  use.alpha = F,
  nsim.blocks = 5,
  path = getwd(),
  append.sims = F,
  block.size = 1000,
  output.name = "model",
  ncores = 1
)
}
\arguments{
\item{model}{A model object built by the main.menu function.}

\item{use.alpha}{Logical. If TRUE the most recent population size change will be exponential. If FALSE sudden demographic changes. Default is FALSE.}

\item{nsim.blocks}{Number of blocks to simulate. The total number of simulations is: nsim.blocks x block.size (x ncores when ncores > 1).}

\item{path}{Path to write the output. By default outputs will be saved in the working directory.}

\item{append.sims}{Logical. If TRUE simulations will be appended in the last output. Default is FALSE.}

\item{block.size}{Simulations are performed in blocks. This argument defines the size of the block in number of simulations, i.e. how many simulations to run per block.
A block of 1000 will work for most cases. Increase the total number of simulations with nsim.blocks argument.}

\item{output.name}{String. The prefix of the output names. Default is "model".}

\item{ncores}{Number of cores for parallel execution. When ncores > 1, separate R worker processes are spawned. Default is 1.}
}
\value{
Writes simulations and parameters to the path directory. msABC outputs a bunch of summary stats by default. They need to be selected a posteriori.
}
\description{
Simulate summary statistics using msABC
}
\note{
This function does not work on Windows systems.
}
\references{
Hudson R.R. (2002) Generating samples under a Wright-Fisher neutral model of genetic variation. Bioinformatics, 18, 337–338.

Pavlidis P., Laurent S., & Stephan W. (2010) msABC: A modification of Hudson’s ms to facilitate multi-locus ABC analysis. Molecular Ecology Resources, 10, 723–727.
}
\author{
Marcelo Gehara
}
