% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/single_pop_demog_test.R
\name{single.pop.demog}
\alias{single.pop.demog}
\title{Simulation of demographic models for single populations.}
\usage{
single.pop.demog(nsims, Ne.prior, time.prior, gene.prior, observed,
  alpha = F, tol = 0.01, nval = 100, do.ABC = F, do.PCA = F,
  CV = F, mod = cbind(c(1, 0.001, 2), c(1, 0.1, 20)), path = path)
}
\arguments{
\item{nsims}{Total number of simulations per model.}

\item{Ne.prior}{Data frame with the prior values for the Ne of each population.}

\item{time.prior}{Data frame with parameter values for the priors of the time of demographic change of each population.}

\item{gene.prior}{Data frame with parameter values for the priors of the mutation rate of each species.}

\item{observed}{Observed summary statistics calculated for the empirical data.}

\item{alpha}{logical. If TRUE all demographic chages are exponential. If FALSE sudden changes. Defaut is FALSE.}

\item{tol}{Tolerance level of the ABC analysis. Ignored if do.ABC = FALSE. Defaut is 0.01.}

\item{nval}{Number of validations for the cross-validation of the ABC. Ignored if do.ABC = FALSE. Default is 100.}

\item{do.ABC}{logical. If TRUE ABC analysis is performed. Default is FALSE.}

\item{do.PCA}{logical. If TRUE PCA of the observed against the simulated data is performed. Default is FALSE.}

\item{CV}{logical. If TRUE cros-validation is performed. Defaut is FALSE}

\item{path}{Path to the directory to write the simulations. Defaut is the working directory.}
}
\description{
Test for constant size, population expantion and bottleneck for single population (non hiersrchical).
}
\details{
This function will take the same inputs used in the codemographic simulations and test for demographic change for each
population separatelly. This test could be usefull to select which populations will be included in the codemographic model and to optimize the prior distributions.
}
